requirements.txt
certifi==2021.5.30
chardet==4.0.0
Django==2.2.4
idna==2.10
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
pytest==6.2.4
mysqlclient
djangorestframework
django_extensions